# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jev Browser Service",
    "description": "Local browser automation for other apps. Submit a job, poll its status, then consume rendered page text and source links. One job executes at a time. **completed** means Jev chose DONE; check result.verification for independent checks. Read [/integration.md](/integration.md) for a complete agent integration guide.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/jobs": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Submit",
        "description": "Queue one goal. Model calls are billable. Do not retry ambiguous submissions blindly.",
        "operationId": "submit_v1_jobs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Jobs",
        "description": "Most recently submitted first. Results remain on disk until explicitly deleted.",
        "operationId": "jobs_v1_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response Jobs V1 Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_v1_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Delete",
        "description": "Delete a finished job and its captured content. Cancel active jobs first.",
        "operationId": "delete_v1_jobs__job_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancel",
        "description": "Request cancellation. Poll until cancelled; already-executed browser actions are not undone.",
        "operationId": "cancel_v1_jobs__job_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BrowserResult": {
        "properties": {
          "agent_status": {
            "type": "string",
            "title": "Agent Status"
          },
          "verification": {
            "type": "string",
            "enum": [
              "passed",
              "failed",
              "not_requested"
            ],
            "title": "Verification"
          },
          "checks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Checks"
          },
          "page": {
            "$ref": "#/components/schemas/PageResult"
          },
          "visited_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Visited Urls"
          },
          "steps": {
            "type": "integer",
            "title": "Steps"
          },
          "elapsed_ms": {
            "type": "integer",
            "title": "Elapsed Ms"
          },
          "actions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Actions"
          },
          "usage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Usage"
          }
        },
        "type": "object",
        "required": [
          "agent_status",
          "verification",
          "checks",
          "page",
          "visited_urls",
          "steps",
          "elapsed_ms",
          "actions"
        ],
        "title": "BrowserResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "blocked",
              "failed",
              "timed_out",
              "cancelled",
              "interrupted"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "request": {
            "$ref": "#/components/schemas/JobRequest"
          },
          "progress": {
            "additionalProperties": true,
            "type": "object",
            "title": "Progress"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrowserResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at",
          "updated_at",
          "request"
        ],
        "title": "Job"
      },
      "JobRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url",
            "description": "Starting HTTP(S) URL. No file or javascript URLs."
          },
          "goal": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Goal",
            "description": "One natural-language browser task."
          },
          "max_steps": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "default": 25
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 600.0,
            "minimum": 5.0,
            "title": "Timeout Seconds",
            "description": "Execution limit, excluding queue time.",
            "default": 120
          },
          "max_text_chars": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 100.0,
            "title": "Max Text Chars",
            "default": 30000
          },
          "capture_screenshots": {
            "type": "boolean",
            "title": "Capture Screenshots",
            "description": "Capture browser previews after decisions.",
            "default": false
          },
          "isolate_browser": {
            "type": "boolean",
            "title": "Isolate Browser",
            "description": "Use a fresh browser context for this job.",
            "default": false
          },
          "verify": {
            "$ref": "#/components/schemas/Verification"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url",
          "goal"
        ],
        "title": "JobRequest"
      },
...
      "PageResult": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "text_truncated": {
            "type": "boolean",
            "title": "Text Truncated"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/SourceLink"
            },
            "type": "array",
            "title": "Links"
          },
          "links_truncated": {
            "type": "boolean",
            "title": "Links Truncated",
            "default": false
          },
          "captured_at": {
            "type": "string",
            "title": "Captured At"
          }
        },
        "type": "object",
        "required": [
          "url",
          "title",
          "text",
          "text_truncated",
          "links",
          "captured_at"
        ],
        "title": "PageResult"
      },
...
      "SourceLink": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "text",
          "url"
        ],
        "title": "SourceLink"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Verification": {
        "properties": {
          "url_contains": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Url Contains"
          },
          "text_contains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Text Contains"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Verification",
        "description": "Optional literal checks on the final rendered document; all must pass."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}